+++ lumigator/frontend/src/stores/modelsStore.ts
@@ -1,6 +1,6 @@
 import { ref } from 'vue'
 import { defineStore } from 'pinia'
-import { modelsService } from '@/sdk/models/modelsService'
+import { modelsService } from '@/sdk/modelsService'
 
 export const useModelStore = defineStore('models', () => {
   const models = ref([])
